Fixed some warnings found by FindBugs. Got rid of the second cell size model and the cell size model interface. Set the maximum new maze size to 64x64, mazes larger than that just don’t  run well. There's still an issue with the stat tracker and different sized mazes.

+++ trunk/src/maze/gui/RobotAnimator.java
@@ -125,7 +125,7 @@ public final class RobotAnimator implements Runnable
                {
                   rotationPercentage += velocity;
                   velocity += acceleration;
-                  rotationPercentage = inc * (4 / inc) / 2;
+                  rotationPercentage = inc * (4 / (double)inc) / 2;
                }
                else
                {
